Refactoring profile view

+++ public/js/views/app/profile.js
@@ -18,9 +18,11 @@ define([
   'views/app/list',
   'views/app/avatar',
 ], function($, Backbone, utils, api, profileTpl, basicInfoTpl, aboutMeTpl, likesTpl, contactTpl, languageTpl, educationTpl, socialTpl, instantTpl, locationTpl, alertTpl, UserProfile, List, avatarView){
+  
   var profileView = Backbone.View.extend({
     el: "#main",
 	markers: {},
+	languagesCount: 0,
 	autoCompleteOptions: { types: ['(cities)'] },
 	events:{
 		"click a#add-language-btn": "addLanguage",
@@ -50,7 +52,6 @@ define([
 			birthYear: '[name=birthYear]',
 			Male: '[name=interestedInM]',
 			Female: '[name=interestedInF]',
-			//CurrentCity, HomeTown, Other locations
 			//About me
 			allAboutYou: '[name=allAboutYou]',
 			mainMission: '[name=mainMission]',
@@ -77,7 +78,8 @@ define([
 	  var sc = this
 	  this.model.fetch({ 
 			success: function(model){
-				console.log("Fetch model:", model, model.bindings)
+				console.log("Fetch model:", model.attributes, model.bindings)
+				sc.languagesCount = model.get("languages").length
 				sc.render()
 			},
 			error: function() { console.log(arguments) }	 	
@@ -85,62 +87,105 @@ define([
 	},
     render: function(){
       $(this.el).html(profileTpl);
-
+	  //Sets tab's contents
 	  $('#basic-info').html(basicInfoTpl)
 	  $('#about-me').html(aboutMeTpl)
 	  $('#likes-info').html(likesTpl)
 	  $('#contact-info').html(contactTpl)
-	  avatarView.render('http://peoplewings-backend.herokuapp.com' + this.model.get("avatar"))
-	  //Clears bindings for languages and educations
+	  avatarView.render(this.model.get("avatar"))
+	  //Clears bindings for x-attributes
 	  this.clearBindings()
 	  //Takes care of languages, intializes bindings, etc
 	  this.initLanguages()
-	  //Takes care of educations, intializes bindings, etc
-	  this.initEducations()
-	  //Takes care of social networks
-	  this.initSocials()
-	  //Takes care of instant Messages
-	  this.initIM()
-	  //Map
+	  //Takes care of educations, socialNets, IMs and otherLocations
+	  this.initLists()
+	  //Canvas assumes a location list exists
 	  this.initCanvas()
-	  
-	  $('.foo').typeahead({
-                                        ajax: {
-                                              url: "http://peoplewings-backend.herokuapp.com/ajax/search/university/",
-                                              timeout: 500,
-                                              //displayField: "name",
-                                              triggerLength: 1,
-                                              method: "get",
-                                              //loadingClass: "loading-circle",
-                                              preDispatch: function (query) {
-                                                  //showLoadingMask(true);
-                                                  return {
-                                                      q: query,
-                                                      otherParam: 123
-                                                  }
-                                              },
-										onselect: function(){
-											console.log(arguments)
-										},
-                                              preProcess: function (data) {
-                                                  //showLoadingMask(false);
-                                                  if (data.code !== 200) {
-                                                      return false;
-                                                  }
-                                                  return data.data;
-                                                              
-                                              }
-                                          }})
-	  
+	  //Inits the upload avatar form
+	  this.initAvatarForm()
     },
+	initAvatarForm: function(){
+		$('a.thumbnail').click(function(e){
+			e.preventDefault();
+	        $('#upload').trigger('click');  
+	    });
+		var sc = this
+		if (window.File && window.FileReader && window.FileList && window.Blob) {
+			function handleFileSelect(evt) {
+	            var files = evt.target.files; // FileList object
+	             console.dir(files)
+				if (files.length > 0) sc.uploadFile(files[0])
+	        }
+			document.getElementById('upload').addEventListener('change', handleFileSelect, false);
+	    } else {
+	            alert('The File APIs are not fully supported in this browser.');
+		}	
+	},
+	uploadFile: function(file){
+		var fd = new FormData();
+	    fd.append("image", file);
+		fd.append("owner", "3");
+		console.log(fd)
+		
+		$.ajax({
+		    url: "http://192.168.1.36:5000/cropper/",
+		    data: fd,
+		    cache: false,
+		    contentType: false,
+		    processData: false,
+		    crossDomain: true,
+		    type: 'POST',
+		    success: function(data){
+		        alert(data);
+		    }
+		});
+
+		//var xhr = new XMLHttpRequest();
+	    
+	    //xhr.addEventListener("load", this.uploadComplete, false);
+	    //xhr.addEventListener("error", this.failCb, false);
+	    //xhr.open("POST", "http://192.168.1.36:5000/cropper/");
+	    //xhr.setRequestHeader("X-CSRFToken", getCookie('csrftoken'));
+	    //    $('.progress').show()
+	    //xhr.send(fd);
+		//$('#crop-modal').modal('show')
+		
+		//require(["jquery.Jcrop"], 
+		//function(){
+		// $('#cropbox').Jcrop()
+		//})
+		/*
+	    var xhr = new XMLHttpRequest();
+	    xhr.upload.addEventListener("progress", uploadProgress, false);
+	    xhr.addEventListener("load", uploadComplete, false);
+	    xhr.addEventListener("error", uploadFailed, false);
+	    xhr.addEventListener("abort", uploadCanceled, false);
+	    xhr.open("POST", "/cropper/");
+	        xhr.setRequestHeader("X-CSRFToken", getCookie('csrftoken'));
+	        $('.progress').show()
+	    xhr.send(fd);	
+		*/
+	},
+	uploadComplete: function(){
+		console.log("Complete!!!")
+	},
 	close: function(){ 
 		this._modelBinder.unbind()
 	},
 	submitProfile: function(e){
 		e.preventDefault(e);
-		console.log('Submit profile ' + e.target.id)
-		console.log(this.model.attributes, this.model.bindings)
-		this.model.save()
+		console.log('Submit profile ' + e.target.id, this.model.attributes, this.model.bindings)
+		this.model.save(this.feedbackResponse)
+	},
+	feedbackResponse: function(response){
+		$('.alert').remove()
+		var tpl
+		if (response.status === true){
+			tpl = _.template(alertTpl, {extraClass: 'alert-success', heading: response.msg })
+		} else {
+			tpl = _.template(alertTpl, {extraClass: 'alert-error', heading: response.msg })
+		}
+		$('#main').prepend(tpl)
 	},
 	initLanguages: function(){
 	  if (this.languagesCount === 0){
@@ -152,10 +197,8 @@ define([
 	  }else this.setLanguages(this.model.get("languages"))
 		
 	  this._modelBinder.bind(this.model, this.el, this.model.bindings)
-	  
 	},
 	setLanguages: function(languages){
-		//console.log(languages)
 		var size = languages.length
 		for (var i = 1; i < size + 1; i++){
 			var tpl = _.template(languageTpl, {index: i})
@@ -171,8 +214,6 @@ define([
 		this.languagesCount += size
 	},
 	addLanguage: function(e){
-		//console.log(this.languagesCount)
-		//if (this.languagesCount == 4) return false
 		e.preventDefault(e);
 		var langs = $('div[id^=select-lang-]').length + 1
 		this.model.set("x_language_" + langs)
@@ -196,37 +237,15 @@ define([
 		this.model.unset("x_level_" + id)
 		this.languagesCount--
 	},
-	clearBindings: function(){
-		for (binding in this.model.bindings){
-			if (binding.indexOf('x_') == 0){
-				delete this.model.bindings[binding]
-			} 
-		}
-	},
-	initEducations: function(){
+	initLists: function(){
 		this.educationList = new List({
 			el: "#education-list", 
 			tpl: educationTpl, 
 			keys: ["institution", "degree"],
 			items: this.model.get("education"),
 			itemId: 'select-study',
-			extraCls: 'foo'
-		})
-		this.educationList.render()
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
-	},
-	addEducation: function(e){
-		e.preventDefault(e);
-		this.educationList.addItem();
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
-		return false
-	},
-	deleteEducation: function(e){
-		console.log("Delete education", e.target.id)
-		this.educationList.deleteItem(e.target.id)
-		return false
-	},
-	initSocials: function(){
+			extraCls: 'autocompleteStudy'
+		}).render()
 		this.socialList = new List({
 			el: "#socialNetwork-list", 
 			tpl: socialTpl, 
@@ -234,21 +253,7 @@ define([
 			items: this.model.get("socialNetworks"),
 			itemId: 'select-social',
 			extraCls: ''
-		})
-		this.socialList.render()
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
-	},
-	addSocial: function(e){
-		e.preventDefault(e);
-		this.socialList.addItem();
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
-		return false
-	},
-	deleteSocial: function(e){
-		this.socialList.deleteItem(e.target.id)
-		return false
-	},
-	initIM: function(){
+		}).render()
 		this.imList = new List({
 			el: "#instantMessage-list", 
 			tpl: instantTpl, 
@@ -256,9 +261,59 @@ define([
 			items: this.model.get("instantMessages"),
 			itemId: 'select-im',
 			extraCls: ''
-		})
-		this.imList.render()
+		}).render()
+		this.locationList = new List({
+			el: "#otherLocations-list", 
+			tpl: locationTpl, 
+			keys: ["name"],
+			items: this.model.get("otherLocations"),
+			itemId: 'select-location',
+			extraCls: ''
+		}).render()
+		
+		this.model.bindings['x_current'] = '[name=current]'
+		this.model.bindings['x_hometown'] = '[name=hometown]'
+		
 		this._modelBinder.bind(this.model, this.el, this.model.bindings)
+		
+		this.setStudyAutocomplete()
+	},
+	setStudyAutocomplete: function(){
+		$('.autocompleteStudy').typeahead({
+      	ajax: {
+              url: api.getServerUrl() + "/ajax/search/university/",
+              timeout: 500,
+              triggerLength: 1,
+              method: "get",
+              preDispatch: function (query) {
+                  return {
+                      q: query,
+                      //otherParam: 123
+                  }
+              },
+			  onselect: function(){
+			  	console.log(arguments)
+			  },
+              preProcess: function (data) {
+                  if (data.code !== 200) {
+                      return false;
+                  }
+                  return data.data;
+              }
+          }
+		})
+	},
+	addEducation: function(e){
+		e.preventDefault(e);
+		this.educationList.addItem();
+		this._modelBinder.bind(this.model, this.el, this.model.bindings)
+		return false
+	},
+	addSocial: function(e){
+		e.preventDefault(e);
+		this.socialList.addItem();
+		this._modelBinder.bind(this.model, this.el, this.model.bindings)
+		return false
 	},
 	addIM: function(e){
 		e.preventDefault(e);
@@ -266,33 +321,36 @@ define([
 		this._modelBinder.bind(this.model, this.el, this.model.bindings)
 		return false
 	},
+	deleteEducation: function(e){
+		if ($("#education-list > div").length === 1) return false
+		this.educationList.deleteItem(e.target.id)
+		return false
+	},
+	
+	deleteSocial: function(e){
+		if ($("#socialNetwork-list > div").length === 1) return false
+		this.socialList.deleteItem(e.target.id)
+		return false
+	},
+	
 	deleteIM: function(e){
+		if ($("#instantMessage-list > div").length === 1) return false
 		this.imList.deleteItem(e.target.id)
 		return false
 	},
 	initCanvas: function(){
 		var sc = this
-		this.bindCities()
-		this.initOtherLocations()
 		require(['async!https://maps.googleapis.com/maps/api/js?key=AIzaSyABBKjubUcAk69Kijktx-s0jcNL1cIjZ98&sensor=false&libraries=places&language=en'], 
 		function(){
-			
-	        var mapcanvas = $(document.createElement('div'));
-	        var latlng = new google.maps.LatLng(0,0);
-	        
-			mapcanvas.attr({ id: 'mapcanvas', style: 'height:300px;margin-left:160px;'}).addClass('span6');
+	        var mapcanvas = $(document.createElement('div')).attr({ id: 'mapcanvas', style: 'height:300px;margin-left:160px;'}).addClass('span6');
 	        $('#otherLocations-group').append(mapcanvas)
         
-	        var myOptions = { zoom: 1, center: latlng, mapTypeControl: false, streetViewControl: false, navigationControlOptions: {style: google.maps.NavigationControlStyle.SMALL}, mapTypeId: google.maps.MapTypeId.ROADMAP }
+	        var myOptions = { zoom: 1, center: new google.maps.LatLng(0,0), mapTypeControl: false, streetViewControl: false, navigationControlOptions: {style: google.maps.NavigationControlStyle.SMALL}, mapTypeId: google.maps.MapTypeId.ROADMAP }
 	        sc.map = new google.maps.Map(document.getElementById("mapcanvas"), myOptions);
-	 		//console.log("Map created:", sc.map)
-			
-			var options = { types: ['(cities)'] };
         
 	        var autoHome = new google.maps.places.Autocomplete(document.getElementById("hometown"), sc.autoCompleteOptions);
 	        var autoCurrent = new google.maps.places.Autocomplete(document.getElementById("currentCity"), sc.autoCompleteOptions);
-			
-			var id_last = $("[name^=name-]:last").attr("id")
+			var id_last = $("[name^=name-]:last").attr("id") // Sets the autocomplete on the last location list field
 	        var autoOther = new google.maps.places.Autocomplete(document.getElementById(id_last), sc.autoCompleteOptions);
 	
 			google.maps.event.addListener(autoHome, 'place_changed', sc.setAutocomplete(autoHome, "hometown", sc.map));   
@@ -304,34 +362,14 @@ define([
 			$("#" + id_last).keypress(function(event) { if ( event.which == 13 ) event.preventDefault() })
 			
 			sc.setInitialMarkers(sc)
-  		    
 		})
-	},
-	bindCities: function(){
-		//Bind lat lon also
-		this.model.bindings['x_current'] = '[name=current]'
-		this.model.bindings['x_hometown'] = '[name=hometown]'
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
-	},
-	initOtherLocations: function(){
-		this.locationList = new List({
-			el: "#otherLocations-list", 
-			tpl: locationTpl, 
-			keys: ["name"],
-			items: this.model.get("otherLocations"),
-			itemId: 'select-location',
-			extraCls: ''
-		})
-		this.locationList.render()
-		this._modelBinder.bind(this.model, this.el, this.model.bindings)
 	},
 	addLocation: function(e){
 		e.preventDefault(e);
 		this.locationList.addItem();
-		var sc = this
 		var id_last = $("[name^=name-]:last").attr("id")
-		var autoOther = new google.maps.places.Autocomplete(document.getElementById(id_last), sc.autoCompleteOptions);
-		google.maps.event.addListener(autoOther, 'place_changed', sc.setAutocomplete(autoOther, "name_" + id_last.split("-", 2)[1], this.map));
+		var autoOther = new google.maps.places.Autocomplete(document.getElementById(id_last), this.autoCompleteOptions);
+		google.maps.event.addListener(autoOther, 'place_changed', this.setAutocomplete(autoOther, "name_" + id_last.split("-", 2)[1], this.map));
 		$("#" + id_last).keypress(function(event) { if ( event.which == 13 ) event.preventDefault() })
 		this._modelBinder.bind(this.model, this.el, this.model.bindings)
 		this.clearMapMarkers()
@@ -339,6 +377,8 @@ define([
 		return false
 	},
 	deleteLocation: function(e){
+		console.log("Locs length", $("#otherLocations-list > div").length)
+		if ($("#otherLocations-list > div").length === 1) return false
 		var sc = this
 		var id = e.target.id.split("-", 3)[2]
 		console.log("DeleteLocation:", id, this.model.get("x_name_" + id).split(",", 1)[0])
@@ -350,51 +390,8 @@ define([
 		this.setInitialMarkers(this)
 		return false
 	},
-	setInitialMarkers: function(sc){
-		var	marker, city
-		city = this.model.get("current")
-		marker = new google.maps.Marker({
-					map: sc.map,
-					position: new google.maps.LatLng(city.lat, city.lon),
-					title: city.name + ", " + city.country,
-					icon: 'img/yellow-marker.png'
-		})
-		this.markers["current"] = marker
-		city = this.model.get("hometown")
-		marker = new google.maps.Marker({
-					map: sc.map,
-					position: new google.maps.LatLng(city.lat, city.lon),
-					title: city.name + ", " + city.country,
-					icon: 'img/blue-marker.png'
-		})
-		this.markers["hometown"] = marker
-		/*for (attr in this.model.attributes){
-				if (attr.indexOf("x_name") == 0){
-					city = this.model.get(attr)
-					console.log("City:", city)
-					marker = new google.maps.Marker({
-							map: sc.map,
-							position: new google.maps.LatLng(city.lat, city.lon),
-							title: city.name + ", " + city.country
-					})
-					this.markers[attr] = marker
-				}
-		}*/
-		var locs = sc.model.get("otherLocations")
-		for ( loc in locs) {
-			city = locs[loc]
-			marker = new google.maps.Marker({
-					map: sc.map,
-					position: new google.maps.LatLng(city.lat, city.lon),
-					title: city.name + ", " + city.country
-			})
-			loc = parseInt(loc) + 1
-			this.markers["name_" + loc] = marker
-		}
-
-	
-	},
 	setAutocomplete: function(auto, field){
+		//console.log("Auto!!!!", field)
 		var sc = this
 		return function(){
    			var place = auto.getPlace();
@@ -420,11 +417,6 @@ define([
 			}
 		}
 	},
-	clearMapMarkers: function(){
-		for (marker in this.markers){
-			this.markers[marker].setMap(null)
-		}
-	},
 	getCityAndCountry: function(address_components){
 		var data = {}
 		var component
@@ -442,7 +434,46 @@ define([
 			}
 		  }
 		return data
-	}
+	},
+	setInitialMarkers: function(sc){
+		var city
+		city = this.model.get("current")
+		this.markers["current"] = new google.maps.Marker({
+					map: sc.map,
+					position: new google.maps.LatLng(city.lat, city.lon),
+					title: city.name + ", " + city.country,
+					icon: 'img/yellow-marker.png'
+		})
+		city = this.model.get("hometown")
+		this.markers["hometown"] = new google.maps.Marker({
+					map: sc.map,
+					position: new google.maps.LatLng(city.lat, city.lon),
+					title: city.name + ", " + city.country,
+					icon: 'img/blue-marker.png'
+		})
+		var locs = sc.model.get("otherLocations")
+		for ( loc in locs) {
+			city = locs[loc]
+			loc = parseInt(loc) + 1
+			this.markers["name_" + loc] = new google.maps.Marker({
+					map: sc.map,
+					position: new google.maps.LatLng(city.lat, city.lon),
+					title: city.name + ", " + city.country
+			})
+		}	
+	},
+	clearMapMarkers: function(){
+		for (marker in this.markers){
+			this.markers[marker].setMap(null)
+		}
+	},
+	clearBindings: function(){
+		for (binding in this.model.bindings){
+			if (binding.indexOf('x_') == 0){
+				delete this.model.bindings[binding]
+			} 
+		}
+	 }
   });
 
   return profileView;
